Add states to beta environment (#751)
These states are scheduled to test the license upload API in the beta
environment.
This adds them to the beta environment so that when the change is merged
into the
main branch the infrastructure will be put in place to allow us to
create app clients for them.

Now that we are frequently making config changes to jurisdictions,
the test environment snapshots for the compact config are causing more
friction than
benefit. Beta and Prod are better candidates for snapshot tests, as they
change less
frequently and we care more about the compacts/jurisdictions we are
letting into
those environments. This replaces the test env snapshots with beta env
snapshots
for tracking compact config.

This is a precursor to #708, #744, #706, #707, and #709


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Updated configuration settings for Ohio and Arkansas jurisdictions to
include the "beta" environment.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->