Fix contract for loading data classes
Realized it would be much more bulletproof to
have the data classes do a full round of serialization/deserialization
if a developer attempts to pass data through the constructor. This
would cause the call to fail if the developer should be loading the data
through the 'load_from_database_record' method when pulling data from
the tables.