Add replay-ability for transaction collection process
Previously, if a transaction history collection job failed to process for
whatever reason (such as a service outage), it was incredibly difficult to run
the processing for that specific time period again. This change leverages the
fact that event bridge sends in the timestamp of the time in which the scheduled
task was triggered.

By going off of this time, rather than generating the timestamp in the lambda,
developers can simply rerun the failed step function execution, and it will
process all the transactions for that same time period. Due to the idempotent
nature of the transaction pk/sk design, we will never see duplicate transaction
records for the same transaction ids.