Remove api gateway request validation for license post endpoint
Normally, we have two layers of request body validation. However, in this case, the API gateway
request validation is insufficient for two core reasons:

1. It doesn't identify the row in which the validation error occurred, making it really
 difficult for state IT staff to triage.
2. It doesn't always specify the field name where the validation error occurred which,
combined with the missing row number, will create a miserable developer experience.

The schema validation performed at the lambda layer provides a much clearer error message
for the caller when validation errors occur, so we are removing the redundant schema
validation at the API gateway layer.