Remove method to calculate activity for license
I realized that we cannot definitively know how long a license record has been continuously active based on the data we collect from the states. This is because the state does not tell us if a license has ever had a period of expiration between renewals. For example
```
2022: license issued
2023: license expired
2024: license renewed (after expired for 1 year)
2025: license uploaded into compact connect with current expiration and issuance date
```
In this case, our system has no visibility into that expiration period, so we cannot know if the license has been continuously active since issued.