Update Password Policy for user pools (#961)
As part of working with OT to fill out the SAQ A form, it listed a
section which sets the following password requirements for admin user
login:

1. passwords minimum of 12 characters
2. passwords use both numbers and letters
3. users cannot use last 4 previous passwords when resetting password.

This configures this in our CDK logic for user pools for staff users, as
they include admins that have access to set Authorize.net credentials.

For practitioners, since they do not have any admin access to payment
credentials, we do not need to enforce any of these according to the SAQ
A form. We set the first two items since this was part of the original
request for this ticket.

Closes #899


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

* **New Features**
* Enhanced password policies now require passwords to include digits and
lowercase letters, with a minimum length of 12 characters.
  * Password history enforcement prevents reuse of the last 4 passwords.
* Temporary passwords and password resets have been updated to meet the
new complexity requirements.

* **Tests**
* Test environments updated to reflect strengthened password complexity
rules.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->